requirements.txt
click==8.0.1
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
nltk==3.6.2
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
regex==2021.4.4
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.61.1
wordcloud==1.8.1
